requirements.txt
# Django Core
Django>=5.2.9,<6.0
django-allauth>=0.57.0
Pillow>=10.0.0

# Django REST Framework (for API)
djangorestframework>=3.14.0

# Celery (for background tasks)
celery>=5.3.0
redis>=5.0.0

# Database (PostgreSQL - optional for now, SQLite works for dev)
# psycopg2-binary>=2.9.9

# Caching
# django-redis>=5.4.0  # Uncomment when Redis is set up

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
# ipython>=8.12.0
# django-debug-toolbar>=4.2.0

# Testing
# pytest>=7.4.0
# pytest-django>=4.7.0
# pytest-cov>=4.1.0

# Monitoring (optional)
# sentry-sdk>=1.38.0
